+++ workbaskets/tests/test_views.py
@@ -2,14 +2,12 @@ import datetime
 import os
 import re
 from unittest.mock import MagicMock
-from unittest.mock import PropertyMock
 from unittest.mock import mock_open
 from unittest.mock import patch
 
 import factory
 import pytest
 from bs4 import BeautifulSoup
-from django.conf import settings
 from django.contrib.auth.models import Permission
 from django.db import IntegrityError
 from django.test.client import RequestFactory
@@ -18,7 +16,6 @@ from django.utils.timezone import localtime
 
 from checks.models import TrackedModelCheck
 from checks.tests.factories import TrackedModelCheckFactory
-from checks.tests.factories import TransactionCheckFactory
 from common.inspect_tap_tasks import CeleryTask
 from common.inspect_tap_tasks import TAPTasks
 from common.models.trackedmodel import TrackedModel
@@ -2835,43 +2832,3 @@ def test_reordering_transactions_bug(valid_user_client, user_workbasket):
         pytest.fail(
             "IntegrityError - New trackedmodel cannot be created after reordering then deleting transactions.",
         )
-
-
-def test_dequeue_reverts_current_version(valid_user, user_workbasket):
-    """Test that when dequeueing a workbasket with 2 updates of the same object
-    that the current version is correctly reverted."""
-    measure_create = factories.MeasureFactory.create()
-    measure_update_1 = measure_create.new_version(
-        workbasket=user_workbasket,
-        update_type=UpdateType.UPDATE,
-    )
-    measure_update_2 = measure_create.new_version(
-        workbasket=user_workbasket,
-        update_type=UpdateType.UPDATE,
-    )
-
-    for tx in user_workbasket.transactions.all():
-        TransactionCheckFactory.create(
-            transaction=tx,
-            successful=True,
-            completed=True,
-        )
-
-    assert measure_create.version_group.current_version == measure_create
-
-    with patch(
-        "workbaskets.models.WorkBasket.unchecked_or_errored_transactions",
-        new_callable=PropertyMock,
-    ) as mock_unchecked:
-        mock_unchecked.return_value = Transaction.objects.none()
-        user_workbasket.queue(
-            valid_user.pk,
-            settings.TRANSACTION_SCHEMA,
-        )
-
-    measure_create.refresh_from_db()
-    assert measure_create.version_group.current_version == measure_update_2
-
-    user_workbasket.dequeue()
-    measure_create.refresh_from_db()
-    assert measure_create.version_group.current_version == measure_create
